- passing result to mainactivity and update mainactivity when there’s data change

+++ InventoryApp/app/src/main/java/datnguyen/com/inventoryapp/MainActivity.java
@@ -1,5 +1,6 @@
 package datnguyen.com.inventoryapp;
 
+import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
@@ -10,6 +11,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.SearchView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -19,6 +21,7 @@ import datnguyen.com.inventoryapp.data.ProductDbHelper;
 import datnguyen.com.inventoryapp.data.Supplier;
 
 import static datnguyen.com.inventoryapp.Constants.EXTRA_PRODUCT_KEY;
+import static datnguyen.com.inventoryapp.Constants.EXTRA_UPDATE_PRODUCT_RESULT_KEY;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -55,52 +58,58 @@ public class MainActivity extends AppCompatActivity {
 		mSharedInstance = this;
 
 		mDbHelper = ProductDbHelper.getDbHelper(getApplicationContext());
-		mDbHelper.deleteAllSuppliers();
-		mDbHelper.insertDummySuppliers();
 
-		mDbHelper.deleteAllProducts();
+		boolean createDummyData = false;
+		if (createDummyData) {
+			mDbHelper.deleteAllSuppliers();
+			mDbHelper.insertDummySuppliers();
 
-		//test suppliers
-		ArrayList<Supplier> listSup = mDbHelper.getAllSuppliers();
-		Log.v(TAG_VIEW, "listSup count: " + listSup.size());
+			mDbHelper.deleteAllProducts();
 
-		// test
-		{
-			Product product = new Product();
-			product.setName("Product 1");
-			product.setPrice(350);
-			product.setQuantity(30);
-			product.setThumnailPath("product1.jpg");
-			product.setSupplierId(listSup.get(0).getId());
+			//test suppliers
+			ArrayList<Supplier> listSup = mDbHelper.getAllSuppliers();
+			Log.v(TAG_VIEW, "listSup count: " + listSup.size());
 
-			// insert to db
-			mDbHelper.insertOrUpdateProduct(product);
+			// test
+			{
+				Product product = new Product();
+				product.setName("Product 1");
+				product.setPrice(350);
+				product.setQuantity(30);
+				product.setThumnailPath("product1.jpg");
+				product.setSupplierId(listSup.get(0).getId());
+
+				// insert to db
+				mDbHelper.insertOrUpdateProduct(product);
+			}
+
+			{
+				Product product = new Product();
+				product.setName("Product 2");
+				product.setPrice(1000);
+				product.setQuantity(5);
+				product.setThumnailPath("product2.jpg");
+				product.setSupplierId(listSup.get(1).getId());
+
+				// insert to db
+				mDbHelper.insertOrUpdateProduct(product);
+			}
+
+			// test
+			{
+				Product product = new Product();
+				product.setName("Product 3");
+				product.setPrice(3500);
+				product.setQuantity(600);
+				product.setThumnailPath("product3.jpg");
+				product.setSupplierId(listSup.get(2).getId());
+
+				// insert to db
+				mDbHelper.insertOrUpdateProduct(product);
+			}
 		}
 
-		{
-			Product product = new Product();
-			product.setName("Product 2");
-			product.setPrice(1000);
-			product.setQuantity(5);
-			product.setThumnailPath("product2.jpg");
-			product.setSupplierId(listSup.get(1).getId());
 
-			// insert to db
-			mDbHelper.insertOrUpdateProduct(product);
-		}
-
-		// test
-		{
-			Product product = new Product();
-			product.setName("Product 3");
-			product.setPrice(3500);
-			product.setQuantity(600);
-			product.setThumnailPath("product3.jpg");
-			product.setSupplierId(listSup.get(2).getId());
-
-			// insert to db
-			mDbHelper.insertOrUpdateProduct(product);
-		}
 
 		productAdapter = new ProductAdapter(productList);
 
@@ -182,16 +191,25 @@ public class MainActivity extends AppCompatActivity {
 
 	@Override
 	protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+		super.onActivityResult(requestCode, resultCode, data);
+
 		Log.v(TAG_VIEW, "onActivityResult requestCode: " + requestCode + " - resultCode: " + resultCode);
 		switch (requestCode) {
 			case REQUEST_CODE_EDIT_PRODUCT:
 			{
-				if (resultCode == RESULT_CODE_EDIT_PRODUCT_SUCCESS) {
-					// update db
-					reloadData();
-				} else {
-					// update error, do nothing
+				if (resultCode == Activity.RESULT_OK) {
+					// check real result from intents
+					int result = data.getExtras().getInt(EXTRA_UPDATE_PRODUCT_RESULT_KEY);
+					if (result == RESULT_CODE_EDIT_PRODUCT_SUCCESS) {
+						// update db
+						reloadData();
+						// show toast to let us know it works
+						Toast.makeText(this, "Data updated!", Toast.LENGTH_SHORT).show();
+					} else {
+						// update error, do nothing
+					}
 				}
+
 			}
 				break;
 		}
